requirements.txt
torch>=1.9.0
torchvision
regex
ftfy
tqdm
huggingface_hub
sentencepiece
protobuf
timm
einops
open_clip_torch
opencv-python
matplotlib
numpy
requests
